Make separate examples for Custom Alert and Custom Bottom Sheet
